\name{run_application}
\alias{run_application}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Run application
}
\description{
Runs the application which solves the POMDP problem : When to stop managing or surveying cryptic threatened species ?
}
\usage{
run_application()
}
%- maybe also 'usage' for other objects documented here.
\details{
Set your parameters, run a simulation or solve the problem
}
\value{
Shiny R app
}
\author{
Luz Pascal
}
\examples{
\dontrun{
run_application()
}
}
